Updates & bug fix
Added checks for both import and deactivate users not found.
Fixed the master list and changed the order of the deactivation so some
users wouldn’t be deactivated only to be reactivated.

Fixed bug in adding/removing users to groups in cloudshell - the call
required a list of users, instead of single - sic added [] to the call